pyproject.toml
[project]
name="mtg_bot"
version = "0.0.1"
requires-python = ">= 3.10"
dependencies = [
    "torch",
    "transformers[torch]",
    "sentence-transformers",
    "datasets",
    "evaluate",
    "pandas",
    "numpy",
    "click",
    "dvc",
    "pytest",
    "mlflow",
    "qdrant-client",
    "fastembed-gpu"
]

[project.scripts]
preprocess = "mtg_bot.cli.preprocess:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mtg_bot*"]
namespaces = false